Fixed Logo RSS, Sitemap, Updated to Font Awesome Free 5.15.2
Chrome and Firefox had the problem that they did not show the icons from the RSS and Sitemap plugin. In Safari everything has always worked. I also updated the font and CSS file from Font Awesome to 5.15.2. The CSS file is all.min.css from Font Awesome. It had to be renamed to fontawesome-all.min.css.